изменено на float

+++ main.py
@@ -65,7 +65,7 @@ def getTicker():
 def get_leverage():
     result = futureAPI.get_leverage("BTC-USD")
     log.info("Get: leverage = {}".format( result ))
-    return int(result['leverage'])
+    return float(result['leverage'])
